Replace filled circles with spiral path
Instead of using two concentric circles to create a filled circle, use a modified spiral path with a smooth outside, so as to save a pen-lift operation.

The path begins and ends (lifts and lowers the pen) in the middle, to avoid pen-lift artifacts.

Two options are given -- to fill with a spiral path that's scaled to the line width or to fill with a constant 2.5 loop spiral (default).